// src/app/models/movies.model.ts
export class Movies {
    id?: string;
    backdrop_path?: string;
    budget?: number;
    genres?: {id: number, name: string};
    homepage?: string;
    original_title?: string;
    popularity?: number;
    poster_path?: string;
    release_date?: string;
    runtime?: number;
    overview?: string;
}